rearranged to make it a simpler project structure
1. test that path/pythonpath is set correctly
2. look for .ini and selinum drivers and fail cleanly
if missing
