+++ .gitignore
@@ -1,3 +1,4 @@
 __pycache__/
 geckodriver.*
 
+credentials-scanner.json
