+++ src/transform/change_list.py
@@ -5,10 +5,10 @@ from typing import Dict, Tuple, List
 from datetime import datetime
 from lxml import html
 
-from code.shared.directory_cache import DirectoryCache
-from code.shared.util import convert_json_to_python, convert_python_to_json
-from code.shared import udatetime
-from code.transform import html_helpers
+from shared.directory_cache import DirectoryCache
+from shared.util import convert_json_to_python, convert_python_to_json
+from shared import udatetime
+from transform import html_helpers
 
 class ChangeItem:
     " status data about a link "
