+++ src/transform/content_table.py
@@ -6,7 +6,7 @@ from lxml import html, etree
 from unidecode import unidecode
 import re
 
-from code.shared.directory_cache import DirectoryCache
+from shared.directory_cache import DirectoryCache
 
 class ContentTable():
     """
