+++ tests/test_git.py
@@ -7,7 +7,7 @@ import os
 sys.path.insert(0, os.path.abspath('.'))
 sys.path.insert(0, os.path.abspath('..'))
 
-from ..code.shared.util_git import pull, push, isbehind
+from code.shared.util_git import pull, push, isbehind
 
 def test_status():
