+++ tests/test_regex.py
